Replace AP Scheduler with Celery 🥦
