+++ security_monkey/reporter.py
@@ -24,11 +24,7 @@
 
 from security_monkey.alerter import Alerter
 from security_monkey.monitors import all_monitors
-from security_monkey.account_manager import get_account_by_name
-from security_monkey import app, db
-from security_monkey.datastore import store_exception
-
-import time
+from security_monkey import app
 
 
 class Reporter(object):
@@ -37,57 +33,6 @@ class Reporter(object):
     def __init__(self, account=None, debug=False):
         self.all_monitors = all_monitors(account, debug)
         self.account_alerter = Alerter(watchers_auditors=self.all_monitors, account=account)
-
-    def run(self, account, interval=None):
-        """Starts the process of watchers -> auditors -> alerters """
-        app.logger.info("Starting work on account {}.".format(account))
-        time1 = time.time()
-        mons = self.get_monitors_to_run(account, interval)
-        watchers_with_changes = set()
-
-        for monitor in mons:
-            app.logger.info("Running slurp {} for {} ({} minutes interval)".format(monitor.watcher.i_am_singular,
-                                                                                   account, interval))
-
-            # Batch logic needs to be handled differently:
-            if monitor.batch_support:
-                from security_monkey.scheduler import batch_logic
-                batch_logic(monitor, monitor.watcher, account, False)
-            else:
-                (items, exception_map) = monitor.watcher.slurp()
-                monitor.watcher.find_changes(items, exception_map)
-                if (len(monitor.watcher.created_items) > 0) or (len(monitor.watcher.changed_items) > 0):
-                    watchers_with_changes.add(monitor.watcher.index)
-                monitor.watcher.save()
-
-        db_account = get_account_by_name(account)
-
-        for monitor in self.all_monitors:
-            # Skip over batched items, since they are done:
-            if monitor.batch_support:
-                continue
-
-            for auditor in monitor.auditors:
-                if auditor.applies_to_account(db_account):
-                    items_to_audit = self.get_items_to_audit(monitor.watcher, auditor, watchers_with_changes)
-                    app.logger.info("Running audit {} for {}".format(
-                                    monitor.watcher.index,
-                                    account))
-
-                    try:
-                        auditor.items = items_to_audit
-                        auditor.audit_objects()
-                        auditor.save_issues()
-                    except Exception as e:
-                        store_exception('reporter-run-auditor', (auditor.index, account), e)
-                        continue
-
-        time2 = time.time()
-        app.logger.info('Run Account %s took %0.1f s' % (account, (time2-time1)))
-
-        self.account_alerter.report()
-
-        db.session.close()
 
     def get_monitors_to_run(self, account, interval=None):
         """
