+++ setup.py
@@ -37,7 +37,8 @@ setup(
                 ('data', ['data/aws_accounts.json'])],
     zip_safe=False,
     install_requires=[
-        'APScheduler==3.5.0',
+        'celery==4.1.0',
+        'celery[redis]==4.1.0',
         'Flask==0.10.1',
         'Flask-Mail==0.9.0',
         'Flask-Migrate==1.3.1',
