+++ supervisor/security_monkey_ui.conf
@@ -1,0 +1,13 @@
+# Copy this file to:
+#    /etc/supervisor/conf.d/security_monkey.conf
+# Restart system supervisor
+#   sudo service supervisor restart
+# Control Startup/Shutdown:
+#   sudo supervisorctl
+
+[program:securitymonkeyui]
+user=www-data
+autostart=true
+autorestart=true
+environment=PYTHONPATH='/usr/local/src/security_monkey/',PATH="/usr/local/src/security_monkey/venv/bin:%(ENV_PATH)s"
+command=/usr/local/src/security_monkey/venv/bin/monkey run_api_server
