+++ supervisor/security_monkey_workers.conf
@@ -1,0 +1,19 @@
+# Copy this file to:
+#    /etc/supervisor/conf.d/security_monkey.conf
+# Restart system supervisor
+#   sudo service supervisor restart
+# Control Startup/Shutdown:
+#   sudo supervisorctl
+
+[program:securitymonkeyworkers]
+user=www-data
+autostart=true
+autorestart=true
+numprocs=1
+directory=/usr/local/src/security_monkey/
+environment=PYTHONPATH='/usr/local/src/security_monkey/',PATH="/usr/local/src/security_monkey/venv/bin:%(ENV_PATH)s"
+startsecs=60
+command=/usr/local/src/security_monkey/venv/bin/celery -A security_monkey.task_scheduler.tasks.CELERY worker
+
+; Causes supervisor to send the termination signal (SIGTERM) to the whole process group.
+stopasgroup=true
